Added codegen for `TriggerOptions` types
