+++ packages/blade-codegen/src/index.ts
@@ -1,4 +1,4 @@
-import { NodeFlags, SyntaxKind, addSyntheticLeadingComment, factory } from 'typescript';
+import { NodeFlags, SyntaxKind, factory } from 'typescript';
 
 import { identifiers } from '@/src/constants/identifiers';
 import {
@@ -29,7 +29,6 @@ import {
   withQueryType,
 } from '@/src/declarations';
 import { importBladeCompilerQueryTypesType } from '@/src/declarations';
-import { generateQueryTypeComment } from '@/src/generators/comment';
 import { createImportDeclaration } from '@/src/generators/import';
 import { generateModelTypes } from '@/src/generators/model';
 import { generateNamespaces } from '@/src/generators/namespaces';
@@ -215,150 +214,19 @@ export const generate = (models: Array<Model>): string => {
               factory.createVariableDeclaration(
                 'use',
                 undefined,
-                factory.createTypeLiteralNode(
-                  models.flatMap((model) => {
-                    const comment = generateQueryTypeComment(model, 'use');
-
-                    const modelSyntaxIdentifier = factory.createIdentifier(
-                      convertToPascalCase(model.slug),
-                    );
-
-                    /**
-                     * @example
-                     * ```ts
-                     * user: ReducedFunction & Syntax.User.Singular.RootQueryCaller & { ... };
-                     * ```
-                     */
-                    const singularProperty = factory.createPropertySignature(
-                      undefined,
-                      model.slug,
-                      undefined,
-                      factory.createIntersectionTypeNode([
-                        factory.createExpressionWithTypeArguments(
-                          identifiers.blade.reducedFunction,
-                          undefined,
-                        ),
-                        factory.createTypeReferenceNode(
-                          factory.createQualifiedName(
-                            factory.createQualifiedName(
-                              factory.createQualifiedName(
-                                identifiers.namespace.syntax.name,
-                                modelSyntaxIdentifier,
-                              ),
-                              identifiers.namespace.syntax.singular,
-                            ),
-                            identifiers.namespace.utils.rootQueryCaller,
-                          ),
-                        ),
-                        factory.createTypeLiteralNode(
-                          Object.entries({
-                            after: identifiers.namespace.utils.afterQuery,
-                            before: identifiers.namespace.utils.beforeQuery,
-                            including: identifiers.namespace.utils.includingQuery,
-                            limitedTo: identifiers.namespace.utils.limitedToQuery,
-                            orderedBy: identifiers.namespace.utils.orderedByQuery,
-                            selecting: identifiers.namespace.utils.selectingQuery,
-                            using: identifiers.namespace.utils.usingQuery,
-                            with: identifiers.namespace.utils.withQuery,
-                          }).map(([instruction, utilIdentifier]) =>
-                            factory.createPropertySignature(
-                              undefined,
-                              instruction,
-                              undefined,
-                              factory.createTypeReferenceNode(
-                                factory.createQualifiedName(
-                                  factory.createQualifiedName(
-                                    factory.createQualifiedName(
-                                      identifiers.namespace.syntax.name,
-                                      modelSyntaxIdentifier,
-                                    ),
-                                    identifiers.namespace.syntax.singular,
-                                  ),
-                                  utilIdentifier,
-                                ),
-                              ),
-                            ),
-                          ),
-                        ),
-                      ]),
-                    );
-
-                    /**
-                     * @example
-                     * ```ts
-                     * users: ReducedFunction & Syntax.User.Plural.RootQueryCaller & { ... };
-                     * ```
-                     */
-                    const pluralProperty = factory.createPropertySignature(
-                      undefined,
-                      model.pluralSlug,
-                      undefined,
-                      factory.createIntersectionTypeNode([
-                        factory.createExpressionWithTypeArguments(
-                          identifiers.blade.reducedFunction,
-                          undefined,
-                        ),
-                        factory.createTypeReferenceNode(
-                          factory.createQualifiedName(
-                            factory.createQualifiedName(
-                              factory.createQualifiedName(
-                                identifiers.namespace.syntax.name,
-                                modelSyntaxIdentifier,
-                              ),
-                              identifiers.namespace.syntax.plural,
-                            ),
-                            identifiers.namespace.utils.rootQueryCaller,
-                          ),
-                        ),
-                        factory.createTypeLiteralNode(
-                          Object.entries({
-                            after: identifiers.namespace.utils.afterQuery,
-                            before: identifiers.namespace.utils.beforeQuery,
-                            including: identifiers.namespace.utils.includingQuery,
-                            limitedTo: identifiers.namespace.utils.limitedToQuery,
-                            orderedBy: identifiers.namespace.utils.orderedByQuery,
-                            selecting: identifiers.namespace.utils.selectingQuery,
-                            using: identifiers.namespace.utils.usingQuery,
-                            with: identifiers.namespace.utils.withQuery,
-                          }).map(([instruction, utilIdentifier]) =>
-                            factory.createPropertySignature(
-                              undefined,
-                              instruction,
-                              undefined,
-                              factory.createTypeReferenceNode(
-                                factory.createQualifiedName(
-                                  factory.createQualifiedName(
-                                    factory.createQualifiedName(
-                                      identifiers.namespace.syntax.name,
-                                      modelSyntaxIdentifier,
-                                    ),
-                                    identifiers.namespace.syntax.plural,
-                                  ),
-                                  utilIdentifier,
-                                ),
-                              ),
-                            ),
-                          ),
-                        ),
-                      ]),
-                    );
-
-                    return [
-                      addSyntheticLeadingComment(
-                        singularProperty,
-                        SyntaxKind.MultiLineCommentTrivia,
-                        comment.singular,
-                        true,
+                factory.createTypeLiteralNode([
+                  factory.createPropertySignature(
+                    undefined,
+                    'add',
+                    undefined,
+                    factory.createTypeReferenceNode(
+                      factory.createQualifiedName(
+                        identifiers.namespace.syntax.name,
+                        identifiers.namespace.syntax.getQuery,
                       ),
-                      addSyntheticLeadingComment(
-                        pluralProperty,
-                        SyntaxKind.MultiLineCommentTrivia,
-                        comment.plural,
-                        true,
-                      ),
-                    ];
-                  }),
-                ),
+                    ),
+                  ),
+                ]),
               ),
             ],
             NodeFlags.Const,
@@ -400,477 +268,33 @@ export const generate = (models: Array<Model>): string => {
                       undefined,
                       'add',
                       undefined,
-                      factory.createTypeLiteralNode(
-                        models.flatMap((model) => {
-                          const comment = generateQueryTypeComment(model, 'add');
-
-                          const modelSyntaxIdentifier = factory.createIdentifier(
-                            convertToPascalCase(model.slug),
-                          );
-
-                          /**
-                           * @example
-                           * ```ts
-                           * user: ReducedFunction & Syntax.User.Singular.RootQueryCaller & { ... };
-                           * ```
-                           */
-                          const singularProperty = factory.createPropertySignature(
-                            undefined,
-                            model.slug,
-                            undefined,
-                            factory.createIntersectionTypeNode([
-                              factory.createExpressionWithTypeArguments(
-                                identifiers.blade.reducedFunction,
-                                undefined,
-                              ),
-                              factory.createTypeReferenceNode(
-                                factory.createQualifiedName(
-                                  factory.createQualifiedName(
-                                    factory.createQualifiedName(
-                                      identifiers.namespace.syntax.name,
-                                      modelSyntaxIdentifier,
-                                    ),
-                                    identifiers.namespace.syntax.singular,
-                                  ),
-                                  identifiers.namespace.utils.rootQueryCallerPromise,
-                                ),
-                              ),
-                              factory.createTypeLiteralNode(
-                                Object.entries({
-                                  after: identifiers.namespace.utils.afterQueryPromise,
-                                  before: identifiers.namespace.utils.beforeQueryPromise,
-                                  including:
-                                    identifiers.namespace.utils.includingQueryPromise,
-                                  limitedTo:
-                                    identifiers.namespace.utils.limitedToQueryPromise,
-                                  orderedBy:
-                                    identifiers.namespace.utils.orderedByQueryPromise,
-                                  selecting:
-                                    identifiers.namespace.utils.selectingQueryPromise,
-                                  to: identifiers.namespace.utils.toQueryPromise,
-                                  using: identifiers.namespace.utils.usingQueryPromise,
-                                  with: identifiers.namespace.utils.withQueryPromise,
-                                }).map(([instruction, utilIdentifier]) =>
-                                  factory.createPropertySignature(
-                                    undefined,
-                                    instruction,
-                                    undefined,
-                                    factory.createTypeReferenceNode(
-                                      factory.createQualifiedName(
-                                        factory.createQualifiedName(
-                                          factory.createQualifiedName(
-                                            identifiers.namespace.syntax.name,
-                                            modelSyntaxIdentifier,
-                                          ),
-                                          identifiers.namespace.syntax.singular,
-                                        ),
-                                        utilIdentifier,
-                                      ),
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ]),
-                          );
-
-                          /**
-                           * @example
-                           * ```ts
-                           * users: ReducedFunction & Syntax.User.Plural.RootQueryCaller & { ... };
-                           * ```
-                           */
-                          const pluralProperty = factory.createPropertySignature(
-                            undefined,
-                            model.pluralSlug,
-                            undefined,
-                            factory.createIntersectionTypeNode([
-                              factory.createExpressionWithTypeArguments(
-                                identifiers.blade.reducedFunction,
-                                undefined,
-                              ),
-                              factory.createTypeReferenceNode(
-                                factory.createQualifiedName(
-                                  factory.createQualifiedName(
-                                    factory.createQualifiedName(
-                                      identifiers.namespace.syntax.name,
-                                      modelSyntaxIdentifier,
-                                    ),
-                                    identifiers.namespace.syntax.plural,
-                                  ),
-                                  identifiers.namespace.utils.rootQueryCallerPromise,
-                                ),
-                              ),
-                              factory.createTypeLiteralNode(
-                                Object.entries({
-                                  after: identifiers.namespace.utils.afterQueryPromise,
-                                  before: identifiers.namespace.utils.beforeQueryPromise,
-                                  including:
-                                    identifiers.namespace.utils.includingQueryPromise,
-                                  limitedTo:
-                                    identifiers.namespace.utils.limitedToQueryPromise,
-                                  orderedBy:
-                                    identifiers.namespace.utils.orderedByQueryPromise,
-                                  selecting:
-                                    identifiers.namespace.utils.selectingQueryPromise,
-                                  to: identifiers.namespace.utils.toQueryPromise,
-                                  using: identifiers.namespace.utils.usingQueryPromise,
-                                  with: identifiers.namespace.utils.withQueryPromise,
-                                }).map(([instruction, utilIdentifier]) =>
-                                  factory.createPropertySignature(
-                                    undefined,
-                                    instruction,
-                                    undefined,
-                                    factory.createTypeReferenceNode(
-                                      factory.createQualifiedName(
-                                        factory.createQualifiedName(
-                                          factory.createQualifiedName(
-                                            identifiers.namespace.syntax.name,
-                                            modelSyntaxIdentifier,
-                                          ),
-                                          identifiers.namespace.syntax.plural,
-                                        ),
-                                        utilIdentifier,
-                                      ),
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ]),
-                          );
-
-                          return [
-                            addSyntheticLeadingComment(
-                              singularProperty,
-                              SyntaxKind.MultiLineCommentTrivia,
-                              comment.singular,
-                              true,
-                            ),
-                            addSyntheticLeadingComment(
-                              pluralProperty,
-                              SyntaxKind.MultiLineCommentTrivia,
-                              comment.plural,
-                              true,
-                            ),
-                          ];
-                        }),
+                      factory.createTypeReferenceNode(
+                        factory.createQualifiedName(
+                          identifiers.namespace.syntax.name,
+                          identifiers.namespace.syntax.addQuery,
+                        ),
                       ),
                     ),
                     factory.createPropertySignature(
                       undefined,
                       'remove',
                       undefined,
-                      factory.createTypeLiteralNode(
-                        models.flatMap((model) => {
-                          const comment = generateQueryTypeComment(model, 'remove');
-
-                          const modelSyntaxIdentifier = factory.createIdentifier(
-                            convertToPascalCase(model.slug),
-                          );
-
-                          /**
-                           * @example
-                           * ```ts
-                           * user: ReducedFunction & Syntax.User.Singular.RootQueryCaller & { ... };
-                           * ```
-                           */
-                          const singularProperty = factory.createPropertySignature(
-                            undefined,
-                            model.slug,
-                            undefined,
-                            factory.createIntersectionTypeNode([
-                              factory.createExpressionWithTypeArguments(
-                                identifiers.blade.reducedFunction,
-                                undefined,
-                              ),
-                              factory.createTypeReferenceNode(
-                                factory.createQualifiedName(
-                                  factory.createQualifiedName(
-                                    factory.createQualifiedName(
-                                      identifiers.namespace.syntax.name,
-                                      modelSyntaxIdentifier,
-                                    ),
-                                    identifiers.namespace.syntax.singular,
-                                  ),
-                                  identifiers.namespace.utils.rootQueryCallerPromise,
-                                ),
-                              ),
-                              factory.createTypeLiteralNode(
-                                Object.entries({
-                                  after: identifiers.namespace.utils.afterQueryPromise,
-                                  before: identifiers.namespace.utils.beforeQueryPromise,
-                                  including:
-                                    identifiers.namespace.utils.includingQueryPromise,
-                                  limitedTo:
-                                    identifiers.namespace.utils.limitedToQueryPromise,
-                                  orderedBy:
-                                    identifiers.namespace.utils.orderedByQueryPromise,
-                                  selecting:
-                                    identifiers.namespace.utils.selectingQueryPromise,
-                                  to: identifiers.namespace.utils.toQueryPromise,
-                                  using: identifiers.namespace.utils.usingQueryPromise,
-                                  with: identifiers.namespace.utils.withQueryPromise,
-                                }).map(([instruction, utilIdentifier]) =>
-                                  factory.createPropertySignature(
-                                    undefined,
-                                    instruction,
-                                    undefined,
-                                    factory.createTypeReferenceNode(
-                                      factory.createQualifiedName(
-                                        factory.createQualifiedName(
-                                          factory.createQualifiedName(
-                                            identifiers.namespace.syntax.name,
-                                            modelSyntaxIdentifier,
-                                          ),
-                                          identifiers.namespace.syntax.singular,
-                                        ),
-                                        utilIdentifier,
-                                      ),
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ]),
-                          );
-
-                          /**
-                           * @example
-                           * ```ts
-                           * users: ReducedFunction & Syntax.User.Plural.RootQueryCaller & { ... };
-                           * ```
-                           */
-                          const pluralProperty = factory.createPropertySignature(
-                            undefined,
-                            model.pluralSlug,
-                            undefined,
-                            factory.createIntersectionTypeNode([
-                              factory.createExpressionWithTypeArguments(
-                                identifiers.blade.reducedFunction,
-                                undefined,
-                              ),
-                              factory.createTypeReferenceNode(
-                                factory.createQualifiedName(
-                                  factory.createQualifiedName(
-                                    factory.createQualifiedName(
-                                      identifiers.namespace.syntax.name,
-                                      modelSyntaxIdentifier,
-                                    ),
-                                    identifiers.namespace.syntax.plural,
-                                  ),
-                                  identifiers.namespace.utils.rootQueryCallerPromise,
-                                ),
-                              ),
-                              factory.createTypeLiteralNode(
-                                Object.entries({
-                                  after: identifiers.namespace.utils.afterQueryPromise,
-                                  before: identifiers.namespace.utils.beforeQueryPromise,
-                                  including:
-                                    identifiers.namespace.utils.includingQueryPromise,
-                                  limitedTo:
-                                    identifiers.namespace.utils.limitedToQueryPromise,
-                                  orderedBy:
-                                    identifiers.namespace.utils.orderedByQueryPromise,
-                                  selecting:
-                                    identifiers.namespace.utils.selectingQueryPromise,
-                                  to: identifiers.namespace.utils.toQueryPromise,
-                                  using: identifiers.namespace.utils.usingQueryPromise,
-                                  with: identifiers.namespace.utils.withQueryPromise,
-                                }).map(([instruction, utilIdentifier]) =>
-                                  factory.createPropertySignature(
-                                    undefined,
-                                    instruction,
-                                    undefined,
-                                    factory.createTypeReferenceNode(
-                                      factory.createQualifiedName(
-                                        factory.createQualifiedName(
-                                          factory.createQualifiedName(
-                                            identifiers.namespace.syntax.name,
-                                            modelSyntaxIdentifier,
-                                          ),
-                                          identifiers.namespace.syntax.plural,
-                                        ),
-                                        utilIdentifier,
-                                      ),
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ]),
-                          );
-
-                          return [
-                            addSyntheticLeadingComment(
-                              singularProperty,
-                              SyntaxKind.MultiLineCommentTrivia,
-                              comment.singular,
-                              true,
-                            ),
-                            addSyntheticLeadingComment(
-                              pluralProperty,
-                              SyntaxKind.MultiLineCommentTrivia,
-                              comment.plural,
-                              true,
-                            ),
-                          ];
-                        }),
+                      factory.createTypeReferenceNode(
+                        factory.createQualifiedName(
+                          identifiers.namespace.syntax.name,
+                          identifiers.namespace.syntax.removeQuery,
+                        ),
                       ),
                     ),
                     factory.createPropertySignature(
                       undefined,
                       'set',
                       undefined,
-                      factory.createTypeLiteralNode(
-                        models.flatMap((model) => {
-                          const comment = generateQueryTypeComment(model, 'set');
-
-                          const modelSyntaxIdentifier = factory.createIdentifier(
-                            convertToPascalCase(model.slug),
-                          );
-
-                          /**
-                           * @example
-                           * ```ts
-                           * user: ReducedFunction & Syntax.User.Singular.RootQueryCaller & { ... };
-                           * ```
-                           */
-                          const singularProperty = factory.createPropertySignature(
-                            undefined,
-                            model.slug,
-                            undefined,
-                            factory.createIntersectionTypeNode([
-                              factory.createExpressionWithTypeArguments(
-                                identifiers.blade.reducedFunction,
-                                undefined,
-                              ),
-                              factory.createTypeReferenceNode(
-                                factory.createQualifiedName(
-                                  factory.createQualifiedName(
-                                    factory.createQualifiedName(
-                                      identifiers.namespace.syntax.name,
-                                      modelSyntaxIdentifier,
-                                    ),
-                                    identifiers.namespace.syntax.singular,
-                                  ),
-                                  identifiers.namespace.utils.rootQueryCallerPromise,
-                                ),
-                              ),
-                              factory.createTypeLiteralNode(
-                                Object.entries({
-                                  after: identifiers.namespace.utils.afterQueryPromise,
-                                  before: identifiers.namespace.utils.beforeQueryPromise,
-                                  including:
-                                    identifiers.namespace.utils.includingQueryPromise,
-                                  limitedTo:
-                                    identifiers.namespace.utils.limitedToQueryPromise,
-                                  orderedBy:
-                                    identifiers.namespace.utils.orderedByQueryPromise,
-                                  selecting:
-                                    identifiers.namespace.utils.selectingQueryPromise,
-                                  to: identifiers.namespace.utils.toQueryPromise,
-                                  using: identifiers.namespace.utils.usingQueryPromise,
-                                  with: identifiers.namespace.utils.withQueryPromise,
-                                }).map(([instruction, utilIdentifier]) =>
-                                  factory.createPropertySignature(
-                                    undefined,
-                                    instruction,
-                                    undefined,
-                                    factory.createTypeReferenceNode(
-                                      factory.createQualifiedName(
-                                        factory.createQualifiedName(
-                                          factory.createQualifiedName(
-                                            identifiers.namespace.syntax.name,
-                                            modelSyntaxIdentifier,
-                                          ),
-                                          identifiers.namespace.syntax.singular,
-                                        ),
-                                        utilIdentifier,
-                                      ),
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ]),
-                          );
-
-                          /**
-                           * @example
-                           * ```ts
-                           * users: ReducedFunction & Syntax.User.Plural.RootQueryCaller & { ... };
-                           * ```
-                           */
-                          const pluralProperty = factory.createPropertySignature(
-                            undefined,
-                            model.pluralSlug,
-                            undefined,
-                            factory.createIntersectionTypeNode([
-                              factory.createExpressionWithTypeArguments(
-                                identifiers.blade.reducedFunction,
-                                undefined,
-                              ),
-                              factory.createTypeReferenceNode(
-                                factory.createQualifiedName(
-                                  factory.createQualifiedName(
-                                    factory.createQualifiedName(
-                                      identifiers.namespace.syntax.name,
-                                      modelSyntaxIdentifier,
-                                    ),
-                                    identifiers.namespace.syntax.plural,
-                                  ),
-                                  identifiers.namespace.utils.rootQueryCallerPromise,
-                                ),
-                              ),
-                              factory.createTypeLiteralNode(
-                                Object.entries({
-                                  after: identifiers.namespace.utils.afterQueryPromise,
-                                  before: identifiers.namespace.utils.beforeQueryPromise,
-                                  including:
-                                    identifiers.namespace.utils.includingQueryPromise,
-                                  limitedTo:
-                                    identifiers.namespace.utils.limitedToQueryPromise,
-                                  orderedBy:
-                                    identifiers.namespace.utils.orderedByQueryPromise,
-                                  selecting:
-                                    identifiers.namespace.utils.selectingQueryPromise,
-                                  to: identifiers.namespace.utils.toQueryPromise,
-                                  using: identifiers.namespace.utils.usingQueryPromise,
-                                  with: identifiers.namespace.utils.withQueryPromise,
-                                }).map(([instruction, utilIdentifier]) =>
-                                  factory.createPropertySignature(
-                                    undefined,
-                                    instruction,
-                                    undefined,
-                                    factory.createTypeReferenceNode(
-                                      factory.createQualifiedName(
-                                        factory.createQualifiedName(
-                                          factory.createQualifiedName(
-                                            identifiers.namespace.syntax.name,
-                                            modelSyntaxIdentifier,
-                                          ),
-                                          identifiers.namespace.syntax.plural,
-                                        ),
-                                        utilIdentifier,
-                                      ),
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ]),
-                          );
-
-                          return [
-                            addSyntheticLeadingComment(
-                              singularProperty,
-                              SyntaxKind.MultiLineCommentTrivia,
-                              comment.singular,
-                              true,
-                            ),
-                            addSyntheticLeadingComment(
-                              pluralProperty,
-                              SyntaxKind.MultiLineCommentTrivia,
-                              comment.plural,
-                              true,
-                            ),
-                          ];
-                        }),
+                      factory.createTypeReferenceNode(
+                        factory.createQualifiedName(
+                          identifiers.namespace.syntax.name,
+                          identifiers.namespace.syntax.setQuery,
+                        ),
                       ),
                     ),
                   ]),
